+++ test/exonum-network/transactions-valid.test.js
@@ -1,10 +1,7 @@
 /* eslint-env node, mocha */
 /* eslint-disable no-unused-expressions */
 
-const {
-  mock, exonumAnchoring, btcdotcomAPI,
-  configBtcDotCom, token, provider
-} = require('../constants').module
+const { mock, exonumAnchoring, configSmartbit, configBtcDotCom, provider } = require('../constants').module
 
 const { cfg1, getFullBlock, getBlocks, getTxs, getExonumTx } = require('../mocks/')
 
@@ -16,25 +13,25 @@ describe('Check anchor transactions valid', function () {
   })
 
   it('when transaction, in correct block and anchored', d => {
-    const anchoring = new exonumAnchoring.Anchoring(configBtcDotCom)
-    const tx = 'b4db78bf1bd164e0417fab25055b1f0e3f7fdad44325a5bf1999d86ab44af2c1'
-    const block = 1688
+    const anchoring = new exonumAnchoring.Anchoring(configSmartbit)
+    const hash = 'e327eb66b3a7df8b3822343bd4233af4148d063368e5003f73adb934d945e9ab'
+    const block = 4302
 
-    mock.onGet(`${btcdotcomAPI}/v3/address/tb1q4mg65jafgx2qgq5ssle7m9v62m5t5tmgv2lqdw6ly5nv4tr8kufq4rj8qz/tx`, {
-      params: { api_key: token, pagesize: 50, page: 1 }
-    }).replyOnce(200, getTxs(20, 1))
+    mock.onGet(/address\/tb1qhcacy66m3sry7lwk29auqsu47ftet70ma7slzpldstyjq39fw2eq9xevnx\/op-returns/, {
+      params: { next: null, limit: 50, sort: 'time', dir: 'asc' }
+    }).replyOnce(200, getTxs(33, 1))
 
-    mock.onGet(`${provider}/api/explorer/v1/blocks/${block}`)
+    mock.onGet(`${provider}/api/explorer/v1/block`, { params: { height: block } })
       .replyOnce(200, getFullBlock(block))
 
     mock.onGet(`${provider}/api/explorer/v1/blocks`, {
-      params: { latest: 2001, count: 312 }
-    }).replyOnce(200, getBlocks(2001, 312))
+      params: { latest: 4400, count: 98 }
+    }).replyOnce(200, getBlocks(4401, 312))
 
-    mock.onGet(`${provider}/api/explorer/v1/transactions/${tx}`)
-      .replyOnce(200, getExonumTx(tx))
+    mock.onGet(`${provider}/api/explorer/v1/transactions`, { params: { hash } })
+      .replyOnce(200, getExonumTx(hash))
 
-    anchoring.txStatus(tx, true)
+    anchoring.txStatus(hash, true)
       .then(d => d.status)
       .should
       .eventually
@@ -43,25 +40,25 @@ describe('Check anchor transactions valid', function () {
   })
 
   it('when transaction, in correct block, but not anchored', d => {
-    const anchoring = new exonumAnchoring.Anchoring(configBtcDotCom)
-    const tx = 'e68a605aa5ce04b7e8c73b4ea46b5f4e2393d82bd50495d3f808315be4619c89'
-    const block = 4002
+    const anchoring = new exonumAnchoring.Anchoring(configSmartbit)
+    const hash = 'e327eb66b3a7df8b3822343bd4233af4148d063368e5003f73adb934d945e9ab'
+    const block = 4302
 
-    mock.onGet(`${btcdotcomAPI}/v3/address/tb1q4mg65jafgx2qgq5ssle7m9v62m5t5tmgv2lqdw6ly5nv4tr8kufq4rj8qz/tx`, {
-      params: { api_key: token, pagesize: 50, page: 1 }
-    }).replyOnce(200, getTxs(5, 1))
+    mock.onGet(/address\/tb1qhcacy66m3sry7lwk29auqsu47ftet70ma7slzpldstyjq39fw2eq9xevnx\/op-returns/, {
+      params: { next: null, limit: 50, sort: 'time', dir: 'asc' }
+    }).replyOnce(200, getTxs(1, 1))
 
-    mock.onGet(`${provider}/api/explorer/v1/blocks/${block}`)
+    mock.onGet(`${provider}/api/explorer/v1/block`, { params: { height: block } })
       .replyOnce(200, getFullBlock(block))
 
     mock.onGet(`${provider}/api/explorer/v1/blocks`, {
-      params: { latest: 5003, count: 1000 }
-    }).replyOnce(200, getBlocks(5003, 1000))
+      params: { latest: 4402, count: 100 }
+    }).replyOnce(200, getBlocks(4403, 100))
 
-    mock.onGet(`${provider}/api/explorer/v1/transactions/${tx}`)
-      .replyOnce(200, getExonumTx(tx))
+    mock.onGet(`${provider}/api/explorer/v1/transactions`, { params: { hash } })
+      .replyOnce(200, getExonumTx(hash))
 
-    anchoring.txStatus(tx, true)
+    anchoring.txStatus(hash, true)
       .then(d => d.status)
       .should
       .eventually
@@ -70,13 +67,13 @@ describe('Check anchor transactions valid', function () {
   })
 
   it('when transaction not commited yet', d => {
-    const anchoring = new exonumAnchoring.Anchoring(configBtcDotCom)
-    const tx = 'c59b07e4bf9c79f487957ee3353dca578495a3284e5145214905c9d6874d393f'
+    const anchoring = new exonumAnchoring.Anchoring(configSmartbit)
+    const hash = 'c59b07e4bf9c79f487957ee3353dca578495a3284e5145214905c9d6874d393f'
 
-    mock.onGet(`${provider}/api/explorer/v1/transactions/${tx}`)
-      .replyOnce(200, getExonumTx(tx))
+    mock.onGet(`${provider}/api/explorer/v1/transactions`, { params: { hash } })
+      .replyOnce(200, getExonumTx(hash))
 
-    anchoring.txStatus(tx, true)
+    anchoring.txStatus(hash, true)
       .then(d => d.status)
       .should
       .eventually