+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.6.0 (Jul 26, 2018)
+
+* Added Exonum 0.9 support. #42 #35
+* Added Smartbit API provider. #42
+* Removed BlockTrail API and BTC.COM API provider. #42
+
 ## 0.5.1 (Jul 20, 2018)
 
 * Add BlockCypher API provider. #27 #28 #29 #31 #32 #33 #34 #36